Add support of shared IFCs
git-svn-id: https://littleims.googlecode.com/svn/trunk@38 b802bcca-024a-11de-88c7-29a3b14d5316

+++ modules/hss/src/main/java/org/cipango/ims/hss/db/SptDao.java
@@ -22,6 +22,6 @@ public interface SptDao extends Dao, ImsDao<SPT>
 {
 	void save(SPT spt);
 	SPT findById(Long id);
-	List<SPT> getSptsByIfc(Long ifcId, Integer groupId);
-	List<Integer> getGroups(Long ifcId);
+	List<SPT> getSptsByIfc(Integer ifcId, Integer groupId);
+	List<Integer> getGroups(Integer ifcId);
 }
